+++ packages/components/virtual-list/src/index.ts
@@ -1,0 +1,5 @@
+export {
+  KpVirtualListComponent,
+  KpVirtualRowDirective,
+  KpVirtualRange,
+} from './virtual-list.component';